+++ src/geneflow/extend/agave_step.py
@@ -209,6 +209,7 @@ class AgaveStep(WorkflowStep):
             # get file list from URI
             file_list = DataManager.list(
                 parsed_uri=uri,
+                inclusive=self._step['map']['inclusive'],
                 globstr=self._step['map']['glob'],
                 agave=self._agave
             )