+++ src/geneflow/extend/slurm_step.py
@@ -197,6 +197,7 @@ class SlurmStep(WorkflowStep):
             # get file list from URI
             file_list = DataManager.list(
                 parsed_uri=uri,
+                inclusive=self._step['map']['inclusive'],
                 globstr=self._step['map']['glob']
             )
             if file_list is False: